+++ src/components/widgets/Layout.jsx
@@ -4,6 +4,16 @@ import { COLOR_SECONDARY } from '../../constants';
 
 // noinspection JSUnresolvedFunction
 const styles = StyleSheet.create({
+  helperView: {
+    backgroundColor: '#ffffff',
+    bottom: 0,
+    height: '50%',
+    left: 0,
+    right: 0,
+    position: 'absolute',
+    width: '100%',
+    zIndex: -1,
+  },
   safeAreaView: {
     flex: 1,
   },
@@ -13,20 +23,7 @@ const Layout = ({ children, backgroundColor = COLOR_SECONDARY, hideHelperViews =
   <SafeAreaView style={{ ...styles.safeAreaView, backgroundColor }}>
     <StatusBar animated backgroundColor={backgroundColor} barStyle={'dark-content'} hidden={statusBarHidden} />
 
-    {!hideHelperViews && (
-      <View
-        style={{
-          width: '100%',
-          height: '50%',
-          backgroundColor: '#ffffff',
-          bottom: 0,
-          left: 0,
-          right: 0,
-          position: 'absolute',
-          zIndex: -1,
-        }}
-      />
-    )}
+    {!hideHelperViews && <View style={styles.helperView} />}
 
     {children}
   </SafeAreaView>