+++ CHANGELOG.md
@@ -1,15 +1,19 @@
 # Changelog
 
 ## 1.3.0
+
 - Added button to load more days in diary (max. 30)
 - Added translations for Arabic, Greek, Japanese, Russian, Sinhala, Turkish, and Ukrainian
+- Preparation for PDF export
 - Bug fixes and improvements
 
 ## 1.2.0
+
 - Notifications added
 - Overview of the last 30 days added
 - Added translations for Spanish, Finnish, French, Italian, Dutch, Polish, Romanian and Chinese
 - Bug fixes and improvements
 
 ## 1.1.0
+
 First version